pyproject.toml
[project]
name = "SomatigoX"
version = "1.0.0"
description = "SomatigoX project is started to create simple agent bot to get data and reports from mongo db."
authors = [{ name = "Amit" }]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["pytest", "pyyaml"]
dependencies = [
    "python-dotenv==1.0.0",
    "pymongo==4.6.0",
    "pandas",      # latest version
    "numpy"        # latest version
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"